# openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Denuncias en Barrios",
        "version": "1.0.0",
        "description": "API REST completa para la gestión de denuncias ciudadanas en barrios. Incluye autenticación, gestión de usuarios, administración de denuncias, panel de administración y dashboard con estadísticas.",
        "contact": {
            "name": "Soporte API",
            "email": "soporte@denunciabarrios.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
            "url": "http://172.208.72.155:3085",
            "description": "Servidor de desarrollo"
        }
    ],
...
    "paths": {
...
        "/admin/personal": {
            "post": {
                "summary": "Registrar nuevo personal municipal",
                "tags": ["Administrador"],
                "security": [{"BearerAuth": []}],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": ["nombreCompleto", "cedula", "cargo", "departamento", "especialidad"],
                                "properties": {
                                    "nombreCompleto": {"type": "string", "example": "Juan Pérez"},
                                    "cedula": {"type": "string", "example": "1234567890"},
                                    "cargo": {"type": "string", "example": "Inspector de Obras"},
                                    "departamento": {
                                        "type": "string",
                                        "enum": ["Agua Potable y Alcantarillado", "Gestión Ambiental", "Obras Públicas y Movilidad", "Control Urbano y Construcciones"],
                                        "example": "Obras Públicas y Movilidad"
                                    },
                                    "especialidad": {"type": "string", "example": "Infraestructura vial"},
                                    "telefono": {"type": "string", "example": "0987654321"},
                                    "email": {"type": "string", "example": "juan.perez@municipio.gob.ec"}
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Personal registrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {"$ref": "#/components/schemas/PersonalMunicipal"}
                            }
                        }
                    },
                    "400": {"description": "Error en parámetros o personal ya existe"},
                    "500": {"description": "Error del servidor"}
                }
            },
            "get": {
                "summary": "Obtener lista de personal municipal",
                "tags": ["Administrador"],
                "security": [{"BearerAuth": []}],
                "parameters": [
                    {
                        "in": "query",
                        "name": "departamento",
                        "schema": {"type": "string"},
                        "description": "Filtrar por departamento"
                    },
                    {
                        "in": "query",
                        "name": "estado",
                        "schema": {"type": "string"},
                        "description": "Filtrar por estado"
                    },
                    {
                        "in": "query",
                        "name": "disponible",
                        "schema": {"type": "boolean"},
                        "description": "Filtrar solo personal activo con pocas asignaciones"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de personal obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {"type": "integer"},
                                        "status": {"type": "string"},
                                        "message": {"type": "string"},
                                        "data": {
                                            "type": "array",
                                            "items": {"$ref": "#/components/schemas/PersonalMunicipal"}
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {"description": "Error del servidor"}
                }
            }
        },
        "/admin/personal/{id}": {
            "get": {
                "summary": "Obtener detalles de un personal específico",
                "tags": ["Administrador"],
                "security": [{"BearerAuth": []}],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {"type": "string"},
                        "description": "ID del personal"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del personal obtenidos exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {"$ref": "#/components/schemas/PersonalMunicipal"}
                            }
                        }
                    },
                    "404": {"description": "Personal no encontrado"},
                    "500": {"description": "Error del servidor"}
                }
            },
            "put": {
                "summary": "Actualizar información de personal",
                "tags": ["Administrador"],
                "security": [{"BearerAuth": []}],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {"type": "string"}
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "cargo": {"type": "string"},
                                    "departamento": {"type": "string"},
                                    "especialidad": {"type": "string"},
                                    "telefono": {"type": "string"},
                                    "email": {"type": "string"},
                                    "estado": {"type": "string", "enum": ["Activo", "Inactivo", "En comisión", "De vacaciones"]}
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Personal actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {"$ref": "#/components/schemas/PersonalMunicipal"}
                            }
                        }
                    },
                    "404": {"description": "Personal no encontrado"},
                    "500": {"description": "Error del servidor"}
                }
            },
            "delete": {
                "summary": "Eliminar (desactivar) personal municipal",
                "tags": ["Administrador"],
                "security": [{"BearerAuth": []}],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {"type": "string"}
                    }
                ],
                "responses": {
                    "200": {"description": "Personal eliminado exitosamente"},
                    "404": {"description": "Personal no encontrado"},
                    "500": {"description": "Error del servidor"}
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Token JWT de autenticación de usuario"
            },
            "AdminAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "auth-admin",
                "description": "Token JWT de autenticación de administrador"
            }
        },
        "schemas": {
...
            "PersonalMunicipal": {
                "type": "object",
                "required": ["nombreCompleto", "cedula", "cargo", "departamento", "especialidad"],
                "properties": {
                    "_id": {
                        "type": "string",
                        "description": "ID único del personal"
                    },
                    "nombreCompleto": {
                        "type": "string",
                        "description": "Nombre completo del personal municipal"
                    },
                    "cedula": {
                        "type": "string",
                        "description": "Número de cédula único"
                    },
                    "cargo": {
                        "type": "string",
                        "description": "Cargo o posición del personal"
                    },
                    "departamento": {
                        "type": "string",
                        "enum": [
                            "Agua Potable y Alcantarillado",
                            "Gestión Ambiental",
                            "Obras Públicas y Movilidad",
                            "Control Urbano y Construcciones"
                        ],
                        "description": "Departamento al que pertenece"
                    },
                    "especialidad": {
                        "type": "string",
                        "description": "Especialidad técnica del personal"
                    },
                    "telefono": {
                        "type": "string",
                        "description": "Número de teléfono de contacto"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "description": "Correo electrónico de contacto"
                    },
                    "estado": {
                        "type": "string",
                        "enum": ["Activo", "Inactivo", "En comisión", "De vacaciones"],
                        "default": "Activo",
                        "description": "Estado actual del personal"
                    },
                    "denunciasAsignadas": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "IDs de denuncias asignadas al personal"
                    },
                    "denunciasResueltas": {
                        "type": "number",
                        "default": 0,
                        "description": "Contador de denuncias resueltas por el personal"
                    },
                    "calificacionPromedio": {
                        "type": "number",
                        "minimum": 0,
                        "maximum": 5,
                        "description": "Calificación promedio del personal"
                    },
                    "fechaIngreso": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Fecha de ingreso al municipio"
                    }
                },
                "example": {
                    "_id": "6123456789abcdef12345679",
                    "nombreCompleto": "Juan Pérez González",
                    "cedula": "1234567890",
                    "cargo": "Inspector de Obras Públicas",
                    "departamento": "Obras Públicas y Movilidad",
                    "especialidad": "Infraestructura vial y bacheo",
                    "telefono": "0987654321",
                    "email": "juan.perez@municipio.gob.ec",
                    "estado": "Activo",
                    "denunciasAsignadas": ["6123456789abcdef12345678"],
                    "denunciasResueltas": 45,
                    "calificacionPromedio": 4.5,
                    "fechaIngreso": "2023-01-15T00:00:00.000Z"
                }
            },
...
        }
    }
}